backend/requirements.txt
django==4.1.2
mysqlclient==2.1.1
djangorestframework==3.14.0
markdown==3.3
django-filter==22.1
drf-spectacular==0.26.0